+++ src/main/java/admin/adminsiteserver/authentication/domain/LoginMember.java
@@ -1,5 +1,7 @@
 package admin.adminsiteserver.authentication.domain;
 
+import admin.adminsiteserver.common.vo.Author;
+import admin.adminsiteserver.common.domain.RoleType;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
@@ -8,7 +10,24 @@ import lombok.RequiredArgsConstructor;
 public class LoginMember {
     private final Long id;
 
+    private final String email;
+
+    private final String studentNumber;
+
+    private final String name;
+
+    private final RoleType roleType;
+
     public static LoginMember from(MemberAdapter memberAdapter) {
-        return new LoginMember(memberAdapter.getId());
+        return new LoginMember(
+                memberAdapter.getId(),
+                memberAdapter.getEmail(),
+                memberAdapter.getStudentNumber(),
+                memberAdapter.getName(),
+                memberAdapter.getRoleType());
+    }
+
+    public Author toAuthor() {
+        return new Author(id, email, studentNumber, name, roleType);
     }
 }
