+++ src/main/java/admin/adminsiteserver/levelup/ui/dto/LevelUpRequest.java
@@ -3,7 +3,7 @@ package admin.adminsiteserver.levelup.ui.dto;
 import admin.adminsiteserver.levelup.domain.LevelUp;
 import admin.adminsiteserver.levelup.exception.NonExistentRoleException;
 import admin.adminsiteserver.member.domain.Member;
-import admin.adminsiteserver.member.domain.RoleType;
+import admin.adminsiteserver.common.domain.RoleType;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
